requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
aiogram==3.6.0
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.1
starlette==0.37.2
jinja2==3.1.4
SQLAlchemy==2.0.30
alembic==1.13.2
psycopg2-binary==2.9.9
